posix: fix pre-existing build errors in stdio.rs and dirent.rs
- stdio.rs: cast file::write() return (isize) to i64 to match
  file_write() return type
- dirent.rs: wrap malloc::free() call in unsafe block (Rust 2024
  edition requires explicit unsafe blocks within unsafe functions)

+++ posix/src/stdio.rs
@@ -334,7 +334,7 @@ fn file_write(f: *mut File, data: *const u8, len: usize) -> i64 {
             file.buf_len = 0;
         }
         file.buf_dir = BUF_DIR_WRITE;
-        let ret = crate::file::write(file.fd, data, len);
+        let ret = crate::file::write(file.fd, data, len) as i64;
         if ret < 0 {
             file.flags |= FLAG_ERR;
         }
